requirements.txt
streamlit
torch
torchvision
opencv-python-headless
segmentation-models-pytorch
Pillow
numpy